solve: solving two type errors: - Argument of type 'BigIntConstructor' is not assignable to parameter of type '(value: BigInt, index: number, array: BigInt[]) => bigint'. - Type 'BigInt' is not assignable to type 'string | number | bigint | boolean'.

+++ crypto/ts/index.ts
@@ -275,7 +275,7 @@ const decrypt = (
 
     const plaintext: Plaintext = ciphertext.data.map(
         (e: BigInt, i: number): BigInt => {
-            return BigInt(e) - BigInt(mimc7.hash(sharedKey, BigInt(ciphertext.iv) + BigInt(i)))
+            return BigInt(e as bigint) - BigInt(mimc7.hash(sharedKey, BigInt(ciphertext.iv as bigint) + BigInt(i)))
         }
     )
